+++ game/race/errors.go
@@ -2,8 +2,10 @@ package race
 
 import (
 	"errors"
+	"time"
 
 	"github.com/rbrabson/goblin/internal/discmsg"
+	"github.com/rbrabson/goblin/internal/format"
 	"golang.org/x/text/language"
 )
 
@@ -12,11 +14,11 @@ var (
 	ErrAlreadyJoinedRace = errors.New("you have already joined the race")
 	ErrBettingHasOpened  = errors.New("betting has opened, so you can't join the race")
 	ErrBettingNotOpened  = errors.New("betting has not opened yet")
-	ErrConfigNotFound    = errors.New("configuration file not found")
-	ErrMemberNotFound    = errors.New("member not found")
-	ErrNoRacersFound     = errors.New("no racers found")
+	ErrConfigNotFound    = errors.New("the configuration file not found")
+	ErrMemberNotFound    = errors.New("the member was not found")
+	ErrNoRacersFound     = errors.New("no racers were found")
 	ErrRaceHasStarted    = errors.New("the race has already started")
-	ErrRacerNotFound     = errors.New("racer not found")
+	ErrRacerNotFound     = errors.New("the racer was not found")
 )
 
 // The maximum number of race members have already joined the race.
@@ -28,4 +30,15 @@ type ErrRaceFull struct {
 func (e ErrRaceFull) Error() string {
 	p := discmsg.GetPrinter(language.AmericanEnglish)
 	return p.Sprintf("you can't join the race, as there are already %d entered into the race", e.MaxNumRacersAllowed)
+}
+
+// The racers are resting, so the user should try again in a certain amount of time.
+type ErrRacersAreResting struct {
+	waitTime time.Duration
+}
+
+// Error returns the error message for ErrRacersAreResting.
+func (e ErrRacersAreResting) Error() string {
+	p := discmsg.GetPrinter(language.AmericanEnglish)
+	return p.Sprintf("The racers are resting. Try again in %s!", format.Duration(e.waitTime))
 }